promis update

+++ Areas/Admin/Controllers/ProductsController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Hosting;
 using System.IO;
 using CoreAspShop.Utility;
 using Microsoft.AspNetCore.Authorization;
+using System;
 
 namespace CoreAspShop.Areas.Admin.Controllers
 {
@@ -54,7 +55,6 @@ namespace CoreAspShop.Areas.Admin.Controllers
             // Проверяем модель на валидность
             if (!ModelState.IsValid)
                 return View(ProductVM);
-
             // Если модель валидна, добавляем информацию в сущности Entity
             await _db.Products.AddAsync(ProductVM.Product);
 
@@ -74,11 +74,22 @@ namespace CoreAspShop.Areas.Admin.Controllers
             // Проверяем, получены ли файлы из формы
             if (files.Count != 0)
             {
+
                 // Комбинируем полный путь к каталогу сохранения
                 string uploadPath = Path.Combine(webRootPath, SD.ImageFolder);
 
                 // Получаем расширение файла
                 var extension = Path.GetExtension(files[0].FileName);
+
+                if (extension != ".jpg" && extension != ".png")
+                {
+                    TempData["SM"] = "Photo have not a true extension.";
+                    var defaultPhotoUploadPath = Path.Combine(webRootPath, SD.ImageFolder + @"\" + SD.DefaultProductImage);
+                    System.IO.File.Copy(defaultPhotoUploadPath, webRootPath + @"\" + SD.ImageFolder + @"\" + ProductVM.Product.Id + ".png");
+                    productFromDb.Image = $"\\{SD.ImageFolder}\\{ProductVM.Product.Id}.png";
+                    await _db.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
 
                 // Сохраняем изображение на сервере
                 using var fs = new FileStream(Path.Combine(uploadPath, ProductVM.Product.Id + extension), FileMode.Create);
@@ -134,22 +145,45 @@ namespace CoreAspShop.Areas.Admin.Controllers
             if (tempProduct is null)
                 return NotFound();
 
-            if(files.Count != 0)
+            if (files.Count != 0)
             {
+
                 var uploadParh = Path.Combine(rootPath, SD.ImageFolder);
 
                 var extension = Path.GetExtension(files[0].FileName);
 
-                var oldExtension = Path.GetExtension(tempProduct.Image);               
+                if (extension != ".jpg" && extension != ".png")
+                {
 
-                if (System.IO.File.Exists(Path.Combine(rootPath, tempProduct.Id + oldExtension)) is true)
-                    System.IO.File.Delete(Path.Combine(rootPath, tempProduct.Id + oldExtension));
-                
-                using var fs = new FileStream(Path.Combine(uploadParh, ProductVM.Product.Id + extension), FileMode.Create);
+                    tempProduct.Name = ProductVM.Product.Name;
+                    tempProduct.Available = ProductVM.Product.Available;
+                    tempProduct.Price = ProductVM.Product.Price;
+                    tempProduct.ShadeColor = ProductVM.Product.ShadeColor;
+                    tempProduct.SpecialTagsId = ProductVM.Product.SpecialTagsId;
+                    tempProduct.ProductTypeId = ProductVM.Product.ProductTypeId;
 
-                await files[0].CopyToAsync(fs);
 
-                tempProduct.Image = $"\\{SD.ImageFolder}\\{ProductVM.Product.Id}{extension}";
+                    _db.Update(tempProduct);
+
+                    await _db.SaveChangesAsync();
+
+                    TempData["SM"] = "Photo have not a true extension.";
+
+                    return RedirectToAction(nameof(Index));
+                }
+                else
+                {
+                    var oldExtension = Path.GetExtension(tempProduct.Image);
+
+                    if (System.IO.File.Exists(Path.Combine(rootPath, tempProduct.Id + oldExtension)) is true)
+                        System.IO.File.Delete(Path.Combine(rootPath, tempProduct.Id + oldExtension));
+
+                    using var fs = new FileStream(Path.Combine(uploadParh, ProductVM.Product.Id + extension), FileMode.Create);
+
+                    await files[0].CopyToAsync(fs);
+
+                    tempProduct.Image = $"\\{SD.ImageFolder}\\{ProductVM.Product.Id}{extension}";
+                }
             }
 
             tempProduct.Name = ProductVM.Product.Name;
@@ -160,7 +194,7 @@ namespace CoreAspShop.Areas.Admin.Controllers
             tempProduct.ProductTypeId = ProductVM.Product.ProductTypeId;
 
 
-            _db.Update(tempProduct);        
+            _db.Update(tempProduct);
 
             await _db.SaveChangesAsync();
